middle-tier/requirements.txt
uvicorn[standard] >= 0.14.0
fastapi >= 0.65.2
sqlalchemy >= 1.4.18
uvloop
python-dotenv >= 0.18.0
PyMySQL >= 1.0.2
pytz
